pyproject.toml
[project]
name = "haexadron"
version = "0.1.0"
description = "Multi-Agent Orchestrator: OpenClaw + deer-flow with security pipeline"
requires-python = ">=3.14"
dependencies = [
    "bitgn-api-connectrpc-python==0.9.0.1.20260405144405+6b39e6872349",
    "bitgn-api-protocolbuffers-python==34.1.0.1.20260405144405+6b39e6872349",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
]

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.uv]
package = false
